requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
python-multipart==0.0.9
requests==2.32.3
opencv-python==4.10.0.84
numpy==2.1.1
python-dotenv==1.0.1
streamlit==1.38.0
google-generativeai==0.8.2
google-genai==0.3.0
